+++ style_bert_vits2/models/infer.py
@@ -6,8 +6,7 @@ from style_bert_vits2.logging import logger
 from style_bert_vits2.models import commons
 from style_bert_vits2.models.models import SynthesizerTrn
 from style_bert_vits2.models.models_jp_extra import SynthesizerTrn as SynthesizerTrnJPExtra
-from style_bert_vits2.text_processing import cleaned_text_to_sequence, extract_bert_feature
-from style_bert_vits2.text_processing.cleaner import clean_text
+from style_bert_vits2.text_processing import clean_text, cleaned_text_to_sequence, extract_bert_feature
 from style_bert_vits2.text_processing.symbols import SYMBOLS
